requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
stripe==7.8.0
bcrypt==4.1.2
requests==2.32.5
gunicorn==21.2.0
python-dotenv==1.0.0
redis==7.1.0
rq==2.6.1
psycopg2-binary==2.9.11
flask-limiter==4.1.1
locust==2.43.2
alembic==1.18.3
APScheduler==3.11.2
reportlab==4.2.0
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0

